day 1 skip cast

+++ src/main/java/not/hub/aoc2021/day1/SonarSweep1.java
@@ -5,11 +5,11 @@ import not.hub.aoc2021.Solver;
 import java.util.List;
 import java.util.stream.IntStream;
 
-public class SonarSweep1 extends Solver<List<Integer>, Integer> {
+public class SonarSweep1 extends Solver<List<Integer>, Long> {
 
     @Override
-    public Integer solve(List<Integer> input) {
-        return (int) IntStream
+    public Long solve(List<Integer> input) {
+        return IntStream
             .range(0, input.size())
             .takeWhile(i -> i < input.size() - 1)
             .filter(i -> input.get(i) < input.get(i + 1))
